requirement.txt
fastapi
uvicorn
sqlalchemy
snowflake-connector-python==3.11.0
passlib
python-jose
pydantic
bcrypt